% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_brm.R
\name{predict_brm}
\alias{predict_brm}
\title{Fitted probabilities from \code{brm} fits}
\usage{
predict_brm(object, x_new = NULL, va_new = NULL, vb_new = NULL, ...)
}
\arguments{
\item{object}{A fitted object from function \code{brm}.}

\item{x_new}{An optional vector of x.}

\item{va_new}{An optional covariate matrix to make predictions with. If omitted, the original matrix va is used.}

\item{vb_new}{An optional covariate matrix to make predictions with. If vb_new is omitted but va_new is not, then vb_new is set to be equal to va_new. If both vb_new and va_new are omitted, then the original matrix vb is used.}

\item{...}{affecting the predictions produced.}
}
\value{
If x_new is omitted, a matrix consisting of fitted probabilities for p0 = P(y=1|x=0,va,vb) and p1 = P(y=1|x=1,va,vb).

If x_new is supplied, a vector consisting of fitted probabilities px = P(y=1|x=x_new,va,vb).
}
\description{
Calculate fitted probabilities from a fitted binary regression model object.
}
