% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print_brm}
\alias{print_brm}
\title{Ancillary function for printing}
\usage{
print_brm(x, ...)
}
\arguments{
\item{x}{a list obtained with the function 'brm'}

\item{...}{additional arguments affecting the output}
}
\description{
Ancillary function for printing
}
